fix: Closes issue #8 (read body)
OVH and other providers remove the contents of /etc/issue in their
OS templates. So we need to ask the user manually to tell us what the
OS is as a Fallback.

In addition we removed the `OsSupported.php` Installer Policy because
its not really needed. We have already determined that the user has a
supported OS.